requirements.txt
SpeechRecognition
google-cloud-speech
google-cloud-texttospeech
pyaudio
spacy
nltk
python-dotenv
requests
google-generativeai
openai
anthropic
flask
flask-cors
gtts
pygame

pypiwin32
soundfile
